+++ common/BUILD
@@ -41,7 +41,6 @@ java_library(
 java_library(
     name = "text_differencer",
     srcs = [
-        "AbstractTextDifferencer.java",
         "LCSTextDifferencer.java",
         "TextDifferencer.java",
     ],
